requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
async-generator==1.10
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
duckduckgo_search==8.1.1
fastapi==0.128.4
fuzzywuzzy==0.18.0
gunicorn==25.0.3
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
Levenshtein==0.27.3
lxml==6.0.2
mypy_extensions==1.1.0
numpy==2.4.2
outcome==1.3.0.post0
packaging==26.0
pandas==3.0.0
primp==0.15.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-Levenshtein==0.27.3
RapidFuzz==3.14.3
requests==2.32.5
selenium==4.40.0
setuptools==81.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.52.1
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
undetected-chromedriver==3.5.5
urllib3==2.6.3
uvicorn==0.40.0
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==16.0
wsproto==1.3.2
zipp==3.23.0
